+++ Sources/Template.swift
@@ -6,7 +6,7 @@ let NSFileNoSuchFileError = 4
 #endif
 
 /// A class representing a template
-public class Template: ExpressibleByStringLiteral {
+open class Template: ExpressibleByStringLiteral {
   let tokens: [Token]
 
   /// Create a template with a template string
@@ -51,7 +51,7 @@ public class Template: ExpressibleByStringLiteral {
   }
 
   /// Render the given template
-  public func render(_ context: Context? = nil) throws -> String {
+  open func render(_ context: Context? = nil) throws -> String {
     let context = context ?? Context()
     let parser = TokenParser(tokens: tokens, namespace: context.namespace)
     let nodes = try parser.parse()